OGL fixes for wxMotif; added Set/GetClientData to wxMenu in wxMSW/wxMotif/wxStubs. Sorry, I may be overwriting changes Robert made due to a file copying error.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@971 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ utils/ogl/samples/ogledit/view.cpp
@@ -77,7 +77,7 @@ void DiagramView::OnDraw(wxDC *dc)
   /* You might use THIS code if you were scaling 
    * graphics of known size to fit on the page. 
    */ 
-  float w, h; 
+  int w, h; 
 
   // We need to adjust for the graphic size, a formula will be added 
   float maxX = 900;
